+++ react/typings/vtex.dreamstore-icons.d.ts
@@ -1,6 +1,0 @@
-declare module 'vtex.dreamstore-icons' {
-  import { ReactElement } from 'react'
-
-  export const IconAssistantSales: ReactElement
-  export const IconProfile: ReactElement
-}