+++ react/typings/vtex.store-icons.d.ts
@@ -1,0 +1,4 @@
+declare module 'vtex.store-icons' {
+  export const IconAssistantSales: ComponentType<any>
+  export const IconProfile: ComponentType<any>
+}