+++ CHANGELOG.md
@@ -21,6 +21,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Icons sizes and viewboxes.
 - Remove left gap on mobile devices.
 
+## [2.7.4] - 2019-03-01
+
+### Changed
+
+- Add snapshot tests.
+- Using `store-icons` instead of `dreamstore-icons`
+
 ## [2.7.3] - 2019-02-14
 
 ## [2.7.2] - 2019-02-14